+++ src/utils/windows/CharsetConverter.cpp
@@ -32,12 +32,12 @@ namespace WINDOWS
 {
 std::string FromW(const wchar_t* str, size_t length)
 {
-  int result = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, str, length, nullptr, 0, nullptr, nullptr);
+  int result = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, str, static_cast<int>(length), nullptr, 0, nullptr, nullptr);
   if (result == 0)
     return std::string();
 
   auto newStr = std::make_unique<char[]>(result);
-  result = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, str, length, newStr.get(), result, nullptr, nullptr);
+  result = WideCharToMultiByte(CP_UTF8, WC_ERR_INVALID_CHARS, str, static_cast<int>(length), newStr.get(), result, nullptr, nullptr);
   if (result == 0)
     return std::string();
 
@@ -51,12 +51,12 @@ std::string FromW(const std::wstring& str)
 
 std::wstring ToW(const char* str, size_t length)
 {
-  int result = MultiByteToWideChar(CP_UTF8, MB_ERR_INVALID_CHARS, str, length, nullptr, 0);
+  int result = MultiByteToWideChar(CP_UTF8, MB_ERR_INVALID_CHARS, str, static_cast<int>(length), nullptr, 0);
   if (result == 0)
     return std::wstring();
 
   auto newStr = std::make_unique<wchar_t[]>(result);
-  result = MultiByteToWideChar(CP_UTF8, MB_ERR_INVALID_CHARS, str, length, newStr.get(), result);
+  result = MultiByteToWideChar(CP_UTF8, MB_ERR_INVALID_CHARS, str, static_cast<int>(length), newStr.get(), result);
 
   if (result == 0)
     return std::wstring();
